Add files via upload
#Bug Fixes

Fix for when a string identifier (",`) was in a comment, the loader was complaining because the identifier was detected as a stray character.

Fix for the inability to call a script function inside an object class. Normally when a function  is called in the object code , the interpreter seatches for the function in the object 
function table, and if the function is not existing then the interpreter is searching in the general function table. The bug was occuri g because the interpreter did not search
the general function table after a search miss.

The function String.ToInteger($error) produced a fatal error in a failed conversion and not simple error (with the description in the $error variable).


#Current known issues

##You can declare function with the same name in an object class and the interpreter will not complain. If you call the function ,
the first one that tis registered in the internal table will be executed. 

##The functions cannot return negative numbers or positive numbers if you prefix them with a prefix e.g +3. This is a loader bug , as it will register (for example) the 
-1 number as an algebraic expression that has to be calculated. When this (faulty) expression is calculated the result is not the -1 number. I will correct this when I have time ,
until then you can return negative values either if you assign the value in a variable e.g $ret = -1 ; return $ret ; OR if you make the return expression a calculated one like
return (0-1)

##The user functions parameters that are members of structures like lists sre being passed as literals and the interpreter complains when the parameter changes in the function body. This is not intended and it will be fixed in the future. To bypass this limitation you can assign the member in a variable before it is passed in the function or
inside the function . For example : 

test($list.another_list)
{
  param1[] =15 ; --> this will make the interpreter to complain
 $p = param1 ;
 $p[] = 15 ; --> this is OK!
} 

or fix this with : 

$p = $list.another_list ;
test($p)
{
 p[] = 15 ; --> OK!
}
 
#Clarification for the indexing of the [Bytes] type.

The [Bytes] type does not return a variable when it is indexed. For example in the situation of $bytes[0]
the returned value is the character that is on the 0 position. This means that you cannot pass as a parameter in a function the character
like test($bytes[0]) as it does not produces a variable. To achieve this make first an assigment like $bt = $byte[0] ; test($bt).
This is intended as we need the [Bytes] to be fast when we manipulate them and the act of internally creating and returning variable objects is slow.