Add files via upload
This script exposes a HTTP object class that can make simple http post and http get request. This object is very limited , it does not handle redirections and can handle only chunked and normal transmition with no compression. It is very possible that some bugs exists but the class is more of a sample than a complete http library.