Bug Fixes
Fix a bug in the [async] functions that caused a crash when the total threads that were created in the life time of the script was 40000. For simplifying the threads management, some algorithms were tweaked and the info that are returned from the  threadInfo($threadId : String):List  function was limited.