Support for password hashing
Introduce three functions that use the monocypher library to create ids and hash passwords.
createRandomId($bytes : Integer):String
Returns a true random id in hexadecimal form. The $bytes parameters
is the length of the string in bytes. For example a $bytes = 6 returns a 12 characters
length string.

hashString($string:String):String
Returns a string with the hash of the given string and the sdalt that was used.
The form of the returned string is : 
$hash=4d8f0fbde1a2a8a5a76a94e36f1d4fb839f28c86f8bc9196714765d9fb741038$salt=06eb726601759de017e1d3ddc1353406

hashStringWithSalt($string:String,$salt:String):String
Returns in hexadecimal form the hash of the $string with the $salt as salt.

Note : The [hashString] and [hashStringWithSalt] can be used for storing passwords 
in a safe form. The comparison of the hashes MUST be done in the hydra script like 
if($hash1 == $hash2)... This is not as secure with a native [compare] function, but its good enough
For better security, a constant delay when the password is verified can be introduced to hinder brute force attacs.