Add a Guard  in the  .CopyBinary
Hydra+ will emit a fatal error if  the $str.CopyBinary($from_indx,$chars_to_copy) has invalid parameters. Previously Hydra+ caused an Access violation if the $chars_to_copy was more than the characters of the string from the $from_indx to the end of the string. A bug was also fixed that caused access violation errors if the $chars_to_copy was more than the characters of the string from the $from_indx to the end of the string.